frontend/requirements.txt
streamlit==1.37.0
requests==2.32.3
python-dotenv==1.0.1
watchdog==6.0.0